+++ R/utils_batch.R
@@ -5,7 +5,7 @@ assert_batches <- function(batches) {
   }
   reps <- unique(map_int(batches, batch_count_reps))
   msg <- paste(
-    "batched targets for tar_map_reps()",
+    "batched tar_rep() targets for tar_rep_map()",
     "must have equal numbers of reps per batch"
   )
   assert_scalar(reps, msg)
@@ -32,9 +32,9 @@ assert_batch <- function(batch, name) {
 #' @export
 assert_batch.default <- function(batch, name) {
   throw_validate(
-    "Invalid tar_map_reps() batch from target ", name,
+    "Invalid tar_rep_map() batch from target ", name,
     ". Upstream batched targets must either be data frames ",
-    "or lists that use iteration = \"list\"."
+    "or lists that use iteration = \"list\" in tar_rep()."
   )
 }
 
@@ -44,9 +44,9 @@ assert_batch.list <- function(batch, name) {
     batch,
     assert_list,
     msg = paste(
-      "Invalid batched list target", name, "for tar_map_reps().",
+      "Invalid batched list target", name, "for tar_rep_map().",
       "Batched list target deps",
-      "need iteration = \"list\" in the target definition."
+      "need iteration = \"list\" in tar_rep()."
     )
   )
   lapply(batch, assert_reps, name = name)
@@ -63,7 +63,7 @@ assert_reps <- function(rep, name) {
     assert_dbl(
       rep[[elt]],
       paste(
-        "in batched target ", name, " supplied to tar_map_reps(), ", elt,
+        "in batched target ", name, " supplied to tar_rep_map(), ", elt,
         ", must be an integer element of upstream batched targets"
       )
     )