+++ NEWS.md
@@ -4,6 +4,10 @@
 
 * Make the `*_raw()` target factories process `command` the same way whether it is an expression or ordinary language object.
 
+## New features
+
+* Add `tar_rep_map()` and `tar_rep_map_raw()` to perform batched computation downstream of `tar_rep()` (#50).
+
 # tarchetypes 0.2.0
 
 * Allow trailing commas in `tar_plan()` (#40, @kendonB).